refactor: AI 코드리뷰 피드백 publish 메서드 제네릭 적용
