+++ mopl-api/src/main/java/io/mopl/api/playlist/domain/PlaylistRepository.java
@@ -1,6 +1,5 @@
-package io.mopl.api.playlist.repository;
+package io.mopl.api.playlist.domain;
 
-import io.mopl.api.playlist.domain.Playlist;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
