+++ mopl-api/src/main/java/io/mopl/api/playlist/service/PlaylistQueryService.java
@@ -4,13 +4,13 @@ import io.mopl.api.common.dto.CursorResponse;
 import io.mopl.api.common.dto.SortDirection;
 import io.mopl.api.content.dto.ContentSummary;
 import io.mopl.api.playlist.domain.Playlist;
+import io.mopl.api.playlist.domain.PlaylistQueryRepository;
+import io.mopl.api.playlist.domain.PlaylistRepository;
 import io.mopl.api.playlist.domain.PlaylistSubscriptionId;
+import io.mopl.api.playlist.domain.PlaylistSubscriptionRepository;
 import io.mopl.api.playlist.dto.PlaylistDto;
 import io.mopl.api.playlist.dto.PlaylistPage;
 import io.mopl.api.playlist.dto.PlaylistSearchRequest;
-import io.mopl.api.playlist.repository.PlaylistQueryRepository;
-import io.mopl.api.playlist.repository.PlaylistRepository;
-import io.mopl.api.playlist.repository.PlaylistSubscriptionRepository;
 import io.mopl.api.playlist.service.loader.PlaylistContentLoader;
 import io.mopl.api.playlist.service.loader.PlaylistOwnerLoader;
 import io.mopl.api.playlist.service.loader.PlaylistSubscriptionLoader;
