+++ mopl-api/src/main/java/io/mopl/api/playlist/service/PlaylistService.java
@@ -4,17 +4,17 @@ import io.mopl.api.common.error.ContentErrorCode;
 import io.mopl.api.content.domain.ContentRepository;
 import io.mopl.api.playlist.domain.Playlist;
 import io.mopl.api.playlist.domain.PlaylistContentId;
+import io.mopl.api.playlist.domain.PlaylistContentRepository;
+import io.mopl.api.playlist.domain.PlaylistRepository;
 import io.mopl.api.playlist.domain.PlaylistSubscription;
 import io.mopl.api.playlist.domain.PlaylistSubscriptionId;
+import io.mopl.api.playlist.domain.PlaylistSubscriptionRepository;
 import io.mopl.api.playlist.dto.PlaylistCreateRequest;
 import io.mopl.api.playlist.dto.PlaylistDto;
 import io.mopl.api.playlist.dto.PlaylistUpdateRequest;
 import io.mopl.api.playlist.event.PlaylistContentAddedInternalEvent;
 import io.mopl.api.playlist.event.PlaylistCreatedInternalEvent;
 import io.mopl.api.playlist.event.PlaylistSubscribedInternalEvent;
-import io.mopl.api.playlist.repository.PlaylistContentRepository;
-import io.mopl.api.playlist.repository.PlaylistRepository;
-import io.mopl.api.playlist.repository.PlaylistSubscriptionRepository;
 import io.mopl.api.user.dto.UserSummary;
 import io.mopl.api.user.service.UserService;
 import io.mopl.core.error.BusinessException;
